% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gc_muhat.R
\name{calculate_gc_muhat}
\alias{calculate_gc_muhat}
\title{Calculate Mean Response Estimates for Each Dose Level}
\usage{
calculate_gc_muhat(data, glm_obj)
}
\arguments{
\item{data}{A data frame containing the \code{dose} column and any other necessary predictors used in the GLM.}

\item{glm_obj}{A fitted GLM object, typically created by the \code{glm} function.
Default is \code{anovaMod}.}
}
\value{
A numeric vector of mean response estimates (\code{mu_hat}) for each dose level.
}
\description{
This function computes the mean response estimates (\code{mu_hat}) for each dose level using a fitted GLM object.
The input data must contain a \code{dose} column.
}
