% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constOptC_ginv.R
\name{constOptC_ginv}
\alias{constOptC_ginv}
\title{Calculate Constrained Optimal Contrast Using Generalized Inverse}
\usage{
constOptC_ginv(mu, Sinv = NULL, placAdj = FALSE, direction)
}
\arguments{
\item{mu}{A numeric vector of means.}

\item{Sinv}{A matrix representing the generalized inverse of the covariance matrix. Default is NULL.}

\item{placAdj}{Logical, whether to adjust for placebo (default is FALSE).}

\item{direction}{Character, either "increasing" or "decreasing" indicating the direction of the effect.}
}
\value{
A numeric vector representing the constrained optimal contrast.
}
\description{
This function calculates the optimal contrast under the constraint that the control and
the active treatment groups have different signs in the contrast.
}
