% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{mcpmodx_oracle}
\alias{mcpmodx_oracle}
\title{Oracle plug-in covariance v^orc (eq. 3.20), for comparison}
\usage{
mcpmodx_oracle(est)
}
\description{
Oracle plug-in covariance v^orc (eq. 3.20), for comparison
}
\keyword{internal}
