% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{mcpmodx_sandwich}
\alias{mcpmodx_sandwich}
\title{Stacked-sandwich covariance of the marginal arm means}
\usage{
mcpmodx_sandwich(est)
}
\arguments{
\item{est}{output of \code{mcpmodx_estimate}.}
}
\value{
a (J+1)x(J+1) matrix V (the covariance of sqrt(n)(mu_hat - mu)).
}
\description{
Stacked-sandwich covariance of the marginal arm means
}
\keyword{internal}
