% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modContr_ginv.R
\name{modContr_ginv}
\alias{modContr_ginv}
\title{Calculate Optimal Contrasts for Multiple Models Using Generalized Inverse}
\usage{
modContr_ginv(means, W = NULL, Sinv = NULL, placAdj = FALSE, type, direction)
}
\arguments{
\item{means}{A matrix of means.}

\item{W}{A weight matrix. Default is NULL.}

\item{Sinv}{A matrix representing the generalized inverse of the covariance matrix. Default is NULL.}

\item{placAdj}{Logical, whether to adjust for placebo (default is FALSE).}

\item{type}{Character, either "unconstrained" or "constrained".}

\item{direction}{Character, direction of the effect ("increasing" or "decreasing").}
}
\value{
A matrix of optimal contrasts.
}
\description{
This function calculates the optimal contrasts for multiple models.
}
