% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optC_ginv.R
\name{optC_ginv}
\alias{optC_ginv}
\title{Calculate Optimal Contrast Using Generalized Inverse}
\usage{
optC_ginv(mu, Sinv = NULL, placAdj = FALSE)
}
\arguments{
\item{mu}{A numeric vector of means.}

\item{Sinv}{A matrix representing the generalized inverse of the covariance matrix. Default is NULL.}

\item{placAdj}{Logical, whether to adjust for placebo (default is FALSE).}
}
\value{
A numeric vector representing the optimal contrast.
}
\description{
This function calculates the optimal contrast for a given mean vector \code{mu}
and the generalized inverse of the covariance matrix \code{Sinv}.
}
