% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optContr_ginv.R
\name{optContr_ginv}
\alias{optContr_ginv}
\title{Calculate Optimal Contrasts and Critical Value Using Generalized Inverse}
\usage{
optContr_ginv(
  models,
  doses,
  w,
  S,
  placAdj = FALSE,
  type = c("unconstrained", "constrained")
)
}
\arguments{
\item{models}{An object of class \code{Mods} representing the dose-response models.}

\item{doses}{A numeric vector of dose levels. Default is taken from \code{models}.}

\item{w}{A weight vector. Default is NULL.}

\item{S}{A covariance matrix. Default is NULL.}

\item{placAdj}{Logical, whether to adjust for placebo (default is FALSE).}

\item{type}{Character, either "unconstrained" or "constrained". Default is "unconstrained".}
}
\value{
A list containing the contrast matrix, mean response matrix, and correlation matrix.
}
\description{
This function calculates optimal contrasts and critical values for specified models and doses.
}
