% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{optimal_contrasts}
\alias{optimal_contrasts}
\title{Optimal contrasts for a set of candidate shapes}
\usage{
optimal_contrasts(D0, Sigma)
}
\description{
Optimal contrasts for a set of candidate shapes
}
\keyword{internal}
