% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCPModX-package.R
\docType{package}
\name{MCPModX-package}
\alias{MCPModX-package}
\title{MCPModX: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cyang728/MCPModX}
  \item Report bugs at \url{https://github.com/cyang728/MCPModX/issues}
}

}
\author{
\strong{Maintainer}: Cheng-Han Yang \email{cheng.han.yang@uth.tmc.edu} (\href{https://orcid.org/0000-0002-4161-3140}{ORCID})

}
\keyword{internal}
